Add ability to inherit debugger providers from global config
This change adds a setup config option to inherit global debugger
adapter configuration instead of having to redefine it in the project
specific config file. Since some adapters need the absolute path to the
binary having it in the project local file might not be portable.

This change adds the inherit_adapters setup option, by default set to
false. Setting it to false will keep the current behavior, which is to
totally ignore the global nvim-dap configuration. Setting it to true,
which should be explicitly done,  will not read the project local
adapters section, instead using the global section.

This change will not break existing config since setup() function is
optional and defaults work to override adapters.

todo: merge global and local adapter sections instead of replacing.